use lowp for colour precision in gles shaders

+++ system/shaders/ftb1/guishader_frag_fonts.glsl
@@ -27,7 +27,7 @@ varying   lowp vec4 m_colour;
 // SM_FONTS shader
 void main ()
 {
-  float a   = m_colour.a * texture2D(m_samp0, m_cord0.xy).a;
+  lowp float a   = m_colour.a * texture2D(m_samp0, m_cord0.xy).a;
   gl_FragColor.rgb   = m_colour.rgb * a;
   gl_FragColor.b   = a;
 }
